Update test cases

+++ cases-py2many/bin_op_no_annotations.jl
@@ -4,7 +4,7 @@ function mult_int_and_int()::Int64
     return a * 2
 end
 
-function mult_float_and_int()::Float64
+function mult_float_and_int()
     a = 2.0
     return a * 2
 end
